+++ src/charm.py
@@ -23,9 +23,6 @@ import ops
 from charms.data_platform_libs.v0.data_interfaces import DatabaseRequires
 from charms.grafana_agent.v0.cos_agent import COSAgentProvider
 from github_runner_manager import constants
-from github_runner_manager.errors import ReconcileError
-from github_runner_manager.manager.runner_manager import FlushMode
-from github_runner_manager.manager.runner_scaler import RunnerScaler
 from github_runner_manager.platform.platform_provider import TokenError
 from github_runner_manager.utilities import set_env_var
 from ops.charm import (
@@ -49,7 +46,6 @@ from charm_state import (
     IMAGE_INTEGRATION_NAME,
     LABELS_CONFIG_NAME,
     PATH_CONFIG_NAME,
-    RECONCILE_INTERVAL_CONFIG_NAME,
     TOKEN_CONFIG_NAME,
     CharmConfigInvalidError,
     CharmState,
@@ -65,8 +61,6 @@ from errors import (
     RunnerManagerServiceError,
     SubprocessError,
 )
-from event_timer import EventTimer, TimerStatusError
-from factories import create_runner_scaler
 from manager_client import GitHubRunnerManagerClient
 
 # The reconcile loop can get stuck in a charm upgrade. Put a timeout so
@@ -158,6 +152,9 @@ def catch_action_errors(
             logger.exception("Issue with charm configuration")
             self.unit.status = BlockedStatus(str(err))
             event.fail(str(err))
+        except RunnerManagerServiceError as err:
+            logger.exception("Failed runner manager request")
+            event.fail(f"Failed runner manager request: {str(err)}")
 
     return func_with_catch_errors
 
@@ -179,8 +176,6 @@ class GithubRunnerCharm(CharmBase):
         self._log_charm_status()
 
         self._grafana_agent = COSAgentProvider(self)
-
-        self._event_timer = EventTimer(self.unit.name)
 
         self._stored.set_default(
             path=self.config[PATH_CONFIG_NAME],  # for detecting changes
@@ -207,16 +202,12 @@ class GithubRunnerCharm(CharmBase):
             self.on[IMAGE_INTEGRATION_NAME].relation_changed,
             self._on_image_relation_changed,
         )
-        self.framework.observe(self.on.reconcile_runners, self._on_reconcile_runners)
         self.framework.observe(self.on.check_runners_action, self._on_check_runners_action)
-        self.framework.observe(self.on.reconcile_runners_action, self._on_reconcile_runners_action)
         self.framework.observe(self.on.flush_runners_action, self._on_flush_runners_action)
         self.framework.observe(self.on.update_status, self._on_update_status)
         self.database = DatabaseRequires(
             self, relation_name="mongodb", database_name=REACTIVE_MQ_DB_NAME
         )
-        self.framework.observe(self.database.on.database_created, self._on_database_created)
-        self.framework.observe(self.database.on.endpoints_changed, self._on_endpoints_changed)
 
         self._manager_client = GitHubRunnerManagerClient(
             host=manager_service.GITHUB_RUNNER_MANAGER_ADDRESS,
@@ -298,34 +289,7 @@ class GithubRunnerCharm(CharmBase):
     @catch_charm_errors
     def _on_start(self, _: StartEvent) -> None:
         """Handle the start of the charm."""
-        state = self._setup_state()
-
-        self.unit.status = MaintenanceStatus("Starting runners")
-        if not self._get_set_image_ready_status():
-            return
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        self._reconcile_openstack_runners(
-            runner_scaler,
-        )
-
-    def _set_reconcile_timer(self) -> None:
-        """Set the timer for regular reconciliation checks."""
-        self._event_timer.ensure_event_timer(
-            event_name="reconcile-runners",
-            interval=int(self.config[RECONCILE_INTERVAL_CONFIG_NAME]),
-            timeout=RECONCILIATION_INTERVAL_TIMEOUT_SECONDS,
-        )
-
-    def _ensure_reconcile_timer_is_active(self) -> None:
-        """Ensure the timer for reconciliation event is active."""
-        try:
-            reconcile_timer_is_active = self._event_timer.is_active(RECONCILE_RUNNERS_EVENT)
-        except TimerStatusError:
-            logger.exception("Failed to check the reconciliation event timer status")
-        else:
-            if not reconcile_timer_is_active:
-                logger.error("Reconciliation event timer is not activated")
-                self._set_reconcile_timer()
+        self._get_set_image_ready_status()
 
     @catch_charm_errors
     def _on_upgrade_charm(self, _: UpgradeCharmEvent) -> None:
@@ -337,58 +301,26 @@ class GithubRunnerCharm(CharmBase):
     def _on_config_changed(self, _: ConfigChangedEvent) -> None:
         """Handle the configuration change."""
         state = self._setup_state()
-        self._set_reconcile_timer()
-        self._setup_service(state)
 
-        flush_and_reconcile = False
+        flush_runners = False
         if state.charm_config.token != self._stored.token:
             self._stored.token = self.config[TOKEN_CONFIG_NAME]
-            flush_and_reconcile = True
+            flush_runners = True
         if self.config[PATH_CONFIG_NAME] != self._stored.path:
             self._stored.path = self.config[PATH_CONFIG_NAME]
-            flush_and_reconcile = True
+            flush_runners = True
         if self.config[LABELS_CONFIG_NAME] != self._stored.labels:
             self._stored.labels = self.config[LABELS_CONFIG_NAME]
-            flush_and_reconcile = True
-
-        state = self._setup_state()
+            flush_runners = True
 
         if not self._get_set_image_ready_status():
             return
-        if flush_and_reconcile:
-            logger.info("Flush and reconcile on config-changed")
-            runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-            runner_scaler.flush(flush_mode=FlushMode.FLUSH_IDLE)
-            self._reconcile_openstack_runners(
-                runner_scaler,
-            )
 
-    @catch_charm_errors
-    def _on_reconcile_runners(self, _: ReconcileRunnersEvent) -> None:
-        """Event handler for reconciling runners."""
-        self._trigger_reconciliation()
+        self._setup_service(state)
 
-    @catch_charm_errors
-    def _on_database_created(self, _: ops.RelationEvent) -> None:
-        """Handle the MongoDB database created event."""
-        self._trigger_reconciliation()
-
-    @catch_charm_errors
-    def _on_endpoints_changed(self, _: ops.RelationEvent) -> None:
-        """Handle the MongoDB endpoints changed event."""
-        self._trigger_reconciliation()
-
-    def _trigger_reconciliation(self) -> None:
-        """Trigger the reconciliation of runners."""
-        self.unit.status = MaintenanceStatus("Reconciling runners")
-        state = self._setup_state()
-
-        if not self._get_set_image_ready_status():
-            return
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        self._reconcile_openstack_runners(
-            runner_scaler,
-        )
+        if flush_runners:
+            logger.info("Flush runners on config-changed")
+            self._manager_client.flush_runner()
 
     @catch_action_errors
     def _on_check_runners_action(self, event: ActionEvent) -> None:
@@ -397,69 +329,17 @@ class GithubRunnerCharm(CharmBase):
         Args:
             event: The event fired on check_runners action.
         """
-        try:
-            info = self._manager_client.check_runner()
-        except RunnerManagerServiceError as err:
-            logger.exception("Failed check runner request")
-            event.fail(f"Failed check runner request: {str(err)}")
-            return
+        info = self._manager_client.check_runner()
         event.set_results(info)
 
     @catch_action_errors
-    def _on_reconcile_runners_action(self, event: ActionEvent) -> None:
-        """Handle the action of reconcile of runner state.
-
-        Args:
-            event: Action event of reconciling the runner.
-        """
-        self.unit.status = MaintenanceStatus("Reconciling runners")
-        state = self._setup_state()
-
-        if not self._get_set_image_ready_status():
-            event.fail("Openstack image not yet provided/ready.")
-            return
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-
-        self.unit.status = MaintenanceStatus("Reconciling runners")
-        try:
-            delta = runner_scaler.reconcile()
-        except ReconcileError:
-            logger.exception(FAILED_TO_RECONCILE_RUNNERS_MSG)
-            self.unit.status = ActiveStatus(ACTIVE_STATUS_RECONCILIATION_FAILED_MSG)
-            event.fail(FAILED_RECONCILE_ACTION_ERR_MSG)
-            return
-
-        self.unit.status = ActiveStatus()
-        event.set_results({"delta": {"virtual-machines": delta}})
-
-    @catch_action_errors
-    def _on_flush_runners_action(self, event: ActionEvent) -> None:
-        """Handle the action of flushing all runner and reconciling afterwards.
-
-        Args:
-            event: Action event of flushing all runners.
-        """
-        state = self._setup_state()
-
-        # Flushing mode not implemented for OpenStack yet.
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        flushed = runner_scaler.flush(flush_mode=FlushMode.FLUSH_IDLE)
-        logger.info("Flushed %s runners", flushed)
-        self.unit.status = MaintenanceStatus("Reconciling runners")
-        try:
-            delta = runner_scaler.reconcile()
-        except ReconcileError:
-            logger.exception(FAILED_TO_RECONCILE_RUNNERS_MSG)
-            self.unit.status = ActiveStatus(ACTIVE_STATUS_RECONCILIATION_FAILED_MSG)
-            event.fail(FAILED_RECONCILE_ACTION_ERR_MSG)
-            return
-        self.unit.status = ActiveStatus()
-        event.set_results({"delta": {"virtual-machines": delta}})
+    def _on_flush_runners_action(self, _: ActionEvent) -> None:
+        """Handle the action of flushing all runner and reconciling afterwards."""
+        self._manager_client.flush_runner()
 
     @catch_charm_errors
     def _on_update_status(self, _: UpdateStatusEvent) -> None:
         """Handle the update of charm status."""
-        self._ensure_reconcile_timer_is_active()
         self._log_juju_processes()
 
     def _setup_service(self, state: CharmState) -> None:
@@ -467,13 +347,17 @@ class GithubRunnerCharm(CharmBase):
 
         Args:
             state: The charm state.
+
+        Raises:
+            RunnerManagerApplicationError: The runner manager service is not ready for requests or
+                has errors.
         """
         try:
             manager_service.setup(state, self.app.name, self.unit.name)
         except RunnerManagerApplicationError:
             logging.exception("Unable to setup the github-runner-manager service")
-            # Not re-raising error for until the github-runner-manager service replaces the
-            # library.
+            raise
+        self._manager_client.wait_till_ready()
 
     @staticmethod
     def _log_juju_processes() -> None:
@@ -513,25 +397,7 @@ class GithubRunnerCharm(CharmBase):
     @catch_charm_errors
     def _on_stop(self, _: StopEvent) -> None:
         """Handle the stopping of the charm."""
-        self._event_timer.disable_event_timer("reconcile-runners")
-        state = self._setup_state()
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        runner_scaler.flush(FlushMode.FLUSH_BUSY)
-
-    def _reconcile_openstack_runners(self, runner_scaler: RunnerScaler) -> None:
-        """Reconcile the current runners state and intended runner state for OpenStack mode.
-
-        Args:
-            runner_scaler: Scaler used to scale the amount of runners.
-        """
-        self.unit.status = MaintenanceStatus("Reconciling runners")
-        try:
-            runner_scaler.reconcile()
-        except ReconcileError:
-            logger.exception(FAILED_TO_RECONCILE_RUNNERS_MSG)
-            self.unit.status = ActiveStatus(ACTIVE_STATUS_RECONCILIATION_FAILED_MSG)
-        else:
-            self.unit.status = ActiveStatus()
+        self._manager_client.flush_runner(busy=True)
 
     def _install_deps(self) -> None:
         """Install dependences for the charm."""
@@ -558,15 +424,10 @@ class GithubRunnerCharm(CharmBase):
     def _on_debug_ssh_relation_changed(self, _: ops.RelationChangedEvent) -> None:
         """Handle debug ssh relation changed event."""
         self.unit.status = MaintenanceStatus("Added debug-ssh relation")
-        state = self._setup_state()
-
         if not self._get_set_image_ready_status():
             return
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        runner_scaler.flush()
-        self._reconcile_openstack_runners(
-            runner_scaler,
-        )
+
+        self._manager_client.flush_runner()
 
     @catch_charm_errors
     def _on_image_relation_joined(self, _: ops.RelationJoinedEvent) -> None:
@@ -582,17 +443,14 @@ class GithubRunnerCharm(CharmBase):
     @catch_charm_errors
     def _on_image_relation_changed(self, _: ops.RelationChangedEvent) -> None:
         """Handle image relation changed event."""
-        state = self._setup_state()
         self.unit.status = MaintenanceStatus("Update image for runners")
-
         if not self._get_set_image_ready_status():
             return
 
-        runner_scaler = create_runner_scaler(state, self.app.name, self.unit.name)
-        runner_scaler.flush(flush_mode=FlushMode.FLUSH_IDLE)
-        self._reconcile_openstack_runners(
-            runner_scaler,
-        )
+        state = self._setup_state()
+        self._setup_service(state)
+
+        self._manager_client.flush_runner()
 
     def _get_set_image_ready_status(self) -> bool:
         """Check if image is ready for Openstack and charm status accordingly.
@@ -607,6 +465,7 @@ class GithubRunnerCharm(CharmBase):
         if not openstack_image.id:
             self.unit.status = WaitingStatus("Waiting for image over integration.")
             return False
+        self.unit.status = ActiveStatus()
         return True
 
     @staticmethod
@@ -652,6 +511,8 @@ def _setup_runner_manager_user() -> None:
         ]
     )
     execute_command(["/usr/bin/chmod", "700", f"/home/{constants.RUNNER_MANAGER_USER}/.ssh"])
+    # Give the user access to write to /var/log
+    execute_command(["/usr/sbin/usermod", "-a", "-G", "syslog", constants.RUNNER_MANAGER_USER])
 
 
 if __name__ == "__main__":