+++ src/manager_service.py
@@ -5,6 +5,7 @@
 
 import json
 import logging
+import os
 import textwrap
 from pathlib import Path
 
@@ -51,7 +52,15 @@ def setup(state: CharmState, app_name: str, unit_name: str) -> None:
         state: The state of the charm.
         app_name: The Juju application name.
         unit_name: The Juju unit.
+
+    Raises:
+        RunnerManagerApplicationStartError: Setup of the runner manager service has failed.
     """
+    try:
+        if systemd.service_running(GITHUB_RUNNER_MANAGER_SERVICE_NAME):
+            systemd.service_stop(GITHUB_RUNNER_MANAGER_SERVICE_NAME)
+    except SystemdError as err:
+        raise RunnerManagerApplicationStartError(_SERVICE_SETUP_ERROR_MESSAGE) from err
     config = create_application_configuration(state, app_name, unit_name)
     config_file = _setup_config_file(config)
     GITHUB_RUNNER_MANAGER_SERVICE_LOG_DIR.mkdir(parents=True, exist_ok=True)
@@ -151,6 +160,7 @@ def _setup_service_file(config_file: Path, log_file: Path) -> None:
         config_file: The configuration file for the service.
         log_file: The file location to store the logs.
     """
+    python_path = Path(os.getcwd()) / "venv"
     service_file_content = textwrap.dedent(
         f"""\
         [Unit]
@@ -161,8 +171,9 @@ def _setup_service_file(config_file: Path, log_file: Path) -> None:
         User={constants.RUNNER_MANAGER_USER}
         Group={constants.RUNNER_MANAGER_GROUP}
         ExecStart=github-runner-manager --config-file {str(config_file)} --host \
-{GITHUB_RUNNER_MANAGER_ADDRESS} --port {GITHUB_RUNNER_MANAGER_PORT}
+{GITHUB_RUNNER_MANAGER_ADDRESS} --port {GITHUB_RUNNER_MANAGER_PORT} --python-path {str(python_path)}
         Restart=on-failure
+        KillMode=process
         StandardOutput=append:{log_file}
         StandardError=append:{log_file}